# 🚀 AMD Gaia NPU + LLMware Integration with Comprehensive UI