--- a/src/common/services/amdGaiaProvider.js
+++ b/src/common/services/amdGaiaProvider.js
@@ -1,0 +1,474 @@
+/**
+ * AMD Gaia Provider Integration
+ * Implements AMDGaiaProvider with DirectML execution provider, batch inference, and memory pool integration
+ */
+
+const EventEmitter = require('events');
+const { DirectMLAcceleratedInference, ONNXRuntimeDirectMLProvider } = require('./directMLProvider');
+const IOBindingManager = require('./ioBindingManager');
+const { SessionPoolManager } = require('./sessionPoolManager');
+const { hardwareConfig } = require('../config/hardwareConfig');
+
+class AMDGaiaProvider extends EventEmitter {
+    constructor(options = {}) {
+        super();
+        
+        this.options = {
+            executionProvider: 'DirectML',
+            deviceType: 'npu',
+            batchSize: 32,
+            precision: 'fp16',
+            enableBatching: true,
+            enableMemoryOptimization: true,
+            maxConcurrentInferences: 4,
+            ...options
+        };
+        
+        this.isInitialized = false;
+        this.capabilities = {
+            directML: false,
+            batchInference: false,
+            memoryOptimization: false,
+            hardwareAcceleration: false
+        };
+        
+        this.stats = {
+            inferencesRun: 0,
+            batchesProcessed: 0,
+            totalInferenceTime: 0,
+            averageInferenceTime: 0,
+            averageBatchSize: 0,
+            memoryUsage: 0,
+            npuUtilization: 0
+        };
+        
+        this.directMLProvider = null;
+        this.onnxProvider = null;
+        this.ioBindingManager = null;
+        this.sessionPoolManager = null;
+        this.memoryPool = options.memoryPool;
+        
+        this.initializeProvider();
+    }
+    
+    async initializeProvider() {
+        try {
+            console.log('[AMD Gaia] Initializing AMD Gaia NPU provider...');
+            
+            // Initialize DirectML provider
+            this.directMLProvider = new DirectMLAcceleratedInference({
+                providers: ['DmlExecutionProvider', 'CPUExecutionProvider'],
+                enableIOBinding: true,
+                enableProfiling: false
+            });
+            
+            // Initialize ONNX Runtime DirectML provider
+            this.onnxProvider = new ONNXRuntimeDirectMLProvider({
+                deviceFilter: 'npu',
+                enableDebugLayer: false
+            });
+            
+            // Initialize IO binding manager
+            this.ioBindingManager = new IOBindingManager(this.memoryPool, {
+                enableZeroCopy: true,
+                enableMemoryReuse: true
+            });
+            
+            // Initialize session pool manager
+            this.sessionPoolManager = new SessionPoolManager({
+                poolOptions: {
+                    minSessions: 1,
+                    maxSessions: this.options.maxConcurrentInferences
+                }
+            });
+            
+            // Check capabilities
+            await this.checkCapabilities();
+            
+            this.isInitialized = true;
+            this.emit('initialized', { capabilities: this.capabilities });
+            
+            console.log('[AMD Gaia] Provider initialized successfully');
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Provider initialization failed:', error);
+            this.emit('initialization-failed', error);
+        }
+    }
+    
+    async checkCapabilities() {
+        try {
+            // Check DirectML support
+            this.capabilities.directML = await this.checkDirectMLSupport();
+            
+            // Check batch inference support
+            this.capabilities.batchInference = this.options.enableBatching;
+            
+            // Check memory optimization support
+            this.capabilities.memoryOptimization = this.options.enableMemoryOptimization && !!this.memoryPool;
+            
+            // Check hardware acceleration
+            this.capabilities.hardwareAcceleration = this.capabilities.directML;
+            
+            console.log('[AMD Gaia] Capabilities:', this.capabilities);
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Capability check failed:', error);
+        }
+    }
+    
+    async checkDirectMLSupport() {
+        try {
+            const health = await this.directMLProvider.checkDeviceHealth();
+            return health.status === 'healthy';
+        } catch (error) {
+            return false;
+        }
+    }
+    
+    /**
+     * Generate embeddings using AMD Gaia NPU
+     */
+    async generateEmbeddings(texts, options = {}) {
+        try {
+            if (!this.isInitialized) {
+                throw new Error('AMD Gaia provider not initialized');
+            }
+            
+            const config = {
+                batchSize: options.batchSize || this.options.batchSize,
+                precision: options.precision || this.options.precision,
+                device: options.device || this.options.deviceType,
+                ...options
+            };
+            
+            // Process in batches
+            const batches = this.createBatches(texts, config.batchSize);
+            const results = [];
+            
+            for (const batch of batches) {
+                const batchResult = await this.processBatch(batch, config);
+                results.push(...batchResult);
+            }
+            
+            this.updateStats(results.length, batches.length);
+            
+            return results;
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Embedding generation failed:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Process a batch of texts for embedding generation
+     */
+    async processBatch(texts, config) {
+        const startTime = Date.now();
+        
+        try {
+            // Prepare inputs for the embedding model
+            const inputs = await this.prepareEmbeddingInputs(texts, config);
+            
+            // Get or create session for embedding model
+            const modelPath = this.getEmbeddingModelPath(config);
+            const sessionWrapper = await this.sessionPoolManager.getSession(modelPath);
+            
+            // Create IO binding for zero-copy operations
+            const bindingInfo = await this.ioBindingManager.createIOBinding(
+                sessionWrapper.session,
+                inputs
+            );
+            
+            // Run inference
+            const outputs = await this.ioBindingManager.runWithBinding(bindingInfo);
+            
+            // Process outputs to extract embeddings
+            const embeddings = this.extractEmbeddings(outputs, texts.length);
+            
+            // Return session to pool
+            this.sessionPoolManager.returnSession(sessionWrapper);
+            
+            const inferenceTime = Date.now() - startTime;
+            this.emit('batch-processed', {
+                batchSize: texts.length,
+                inferenceTime,
+                throughput: texts.length / (inferenceTime / 1000)
+            });
+            
+            return embeddings;
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Batch processing failed:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Prepare inputs for embedding model
+     */
+    async prepareEmbeddingInputs(texts, config) {
+        try {
+            // Tokenize texts (simplified - would use actual tokenizer)
+            const tokenized = texts.map(text => this.tokenizeText(text, config));
+            
+            // Create input tensors
+            const maxLength = Math.max(...tokenized.map(tokens => tokens.length));
+            const batchSize = texts.length;
+            
+            // Pad sequences to same length
+            const inputIds = new Int32Array(batchSize * maxLength);
+            const attentionMask = new Int32Array(batchSize * maxLength);
+            
+            for (let i = 0; i < batchSize; i++) {
+                const tokens = tokenized[i];
+                for (let j = 0; j < maxLength; j++) {
+                    const idx = i * maxLength + j;
+                    if (j < tokens.length) {
+                        inputIds[idx] = tokens[j];
+                        attentionMask[idx] = 1;
+                    } else {
+                        inputIds[idx] = 0; // Padding token
+                        attentionMask[idx] = 0;
+                    }
+                }
+            }
+            
+            return {
+                input_ids: {
+                    data: inputIds,
+                    dims: [batchSize, maxLength],
+                    type: 'int32'
+                },
+                attention_mask: {
+                    data: attentionMask,
+                    dims: [batchSize, maxLength],
+                    type: 'int32'
+                }
+            };
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Input preparation failed:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Simple tokenization (would use proper tokenizer in production)
+     */
+    tokenizeText(text, config) {
+        // Simplified tokenization - split by spaces and convert to IDs
+        const tokens = text.toLowerCase().split(/\s+/).filter(token => token.length > 0);
+        
+        // Convert to token IDs (simplified mapping)
+        return tokens.map(token => {
+            let hash = 0;
+            for (let i = 0; i < token.length; i++) {
+                const char = token.charCodeAt(i);
+                hash = ((hash << 5) - hash) + char;
+                hash = hash & hash; // Convert to 32-bit integer
+            }
+            return Math.abs(hash) % 30000 + 1; // Vocabulary size simulation
+        });
+    }
+    
+    /**
+     * Extract embeddings from model outputs
+     */
+    extractEmbeddings(outputs, batchSize) {
+        try {
+            // Assuming the model outputs embeddings in 'last_hidden_state' or similar
+            const embeddingOutput = outputs.last_hidden_state || outputs.embeddings || Object.values(outputs)[0];
+            
+            if (!embeddingOutput) {
+                throw new Error('No embedding output found');
+            }
+            
+            // Extract embeddings (typically mean pooling of token embeddings)
+            const embeddings = [];
+            const embeddingDim = embeddingOutput.dims[embeddingOutput.dims.length - 1];
+            const seqLength = embeddingOutput.dims[1];
+            
+            for (let i = 0; i < batchSize; i++) {
+                const embedding = new Float32Array(embeddingDim);
+                
+                // Mean pooling across sequence length
+                for (let j = 0; j < embeddingDim; j++) {
+                    let sum = 0;
+                    for (let k = 0; k < seqLength; k++) {
+                        const idx = i * seqLength * embeddingDim + k * embeddingDim + j;
+                        sum += embeddingOutput.data[idx];
+                    }
+                    embedding[j] = sum / seqLength;
+                }
+                
+                embeddings.push(embedding);
+            }
+            
+            return embeddings;
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Embedding extraction failed:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Get embedding model path based on configuration
+     */
+    getEmbeddingModelPath(config) {
+        // Return appropriate model path based on precision and other config
+        const modelName = config.precision === 'fp16' ? 'embedding_model_fp16.onnx' : 'embedding_model_fp32.onnx';
+        return require('path').join(process.cwd(), 'models', modelName);
+    }
+    
+    /**
+     * Create batches from input texts
+     */
+    createBatches(texts, batchSize) {
+        const batches = [];
+        for (let i = 0; i < texts.length; i += batchSize) {
+            batches.push(texts.slice(i, i + batchSize));
+        }
+        return batches;
+    }
+    
+    /**
+     * Run inference on AMD Gaia NPU
+     */
+    async runInference(modelPath, inputs, options = {}) {
+        try {
+            if (!this.isInitialized) {
+                throw new Error('AMD Gaia provider not initialized');
+            }
+            
+            const sessionWrapper = await this.sessionPoolManager.getSession(modelPath);
+            
+            // Use DirectML provider for inference
+            const result = await this.directMLProvider.runSingleInference(
+                sessionWrapper.session,
+                inputs,
+                options
+            );
+            
+            this.sessionPoolManager.returnSession(sessionWrapper);
+            
+            this.stats.inferencesRun++;
+            
+            return result;
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Inference failed:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Update provider statistics
+     */
+    updateStats(inferenceCount, batchCount) {
+        this.stats.inferencesRun += inferenceCount;
+        this.stats.batchesProcessed += batchCount;
+        
+        // Update averages
+        if (this.stats.batchesProcessed > 0) {
+            this.stats.averageBatchSize = this.stats.inferencesRun / this.stats.batchesProcessed;
+        }
+    }
+    
+    /**
+     * Get provider statistics
+     */
+    getStats() {
+        return {
+            ...this.stats,
+            capabilities: this.capabilities,
+            isInitialized: this.isInitialized,
+            sessionPoolStats: this.sessionPoolManager?.getGlobalStats(),
+            ioBindingStats: this.ioBindingManager?.getStats(),
+            directMLStats: this.directMLProvider?.getStats()
+        };
+    }
+    
+    /**
+     * Get hardware utilization metrics
+     */
+    async getHardwareMetrics() {
+        try {
+            // Get NPU utilization (would interface with actual hardware monitoring)
+            const metrics = {
+                npuUtilization: this.stats.npuUtilization,
+                memoryUsage: this.stats.memoryUsage,
+                powerConsumption: 0, // Would get from hardware
+                temperature: 0, // Would get from hardware
+                throughput: this.calculateThroughput()
+            };
+            
+            return metrics;
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Hardware metrics failed:', error);
+            return null;
+        }
+    }
+    
+    calculateThroughput() {
+        if (this.stats.totalInferenceTime === 0) return 0;
+        return (this.stats.inferencesRun / (this.stats.totalInferenceTime / 1000));
+    }
+    
+    /**
+     * Optimize model for AMD Gaia NPU
+     */
+    async optimizeModel(modelPath, outputPath, options = {}) {
+        try {
+            const config = hardwareConfig.getConfig('amd-gaia-npu');
+            
+            return await this.directMLProvider.optimizeModelForDirectML(
+                modelPath,
+                outputPath,
+                {
+                    ...config.sessionOptions,
+                    ...options
+                }
+            );
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Model optimization failed:', error);
+            throw error;
+        }
+    }
+    
+    /**
+     * Dispose of the provider
+     */
+    async dispose() {
+        try {
+            if (this.sessionPoolManager) {
+                await this.sessionPoolManager.dispose();
+            }
+            
+            if (this.ioBindingManager) {
+                this.ioBindingManager.dispose();
+            }
+            
+            if (this.directMLProvider) {
+                await this.directMLProvider.dispose();
+            }
+            
+            if (this.onnxProvider) {
+                await this.onnxProvider.dispose();
+            }
+            
+            this.removeAllListeners();
+            
+            console.log('[AMD Gaia] Provider disposed');
+            
+        } catch (error) {
+            console.error('[AMD Gaia] Disposal failed:', error);
+        }
+    }
+}
+
+module.exports = AMDGaiaProvider;