… Tasks_2025-07-05T23-26-57.md
… docs/AMD_Gaia_LLMware_Integration_Guide.md
… docs/Configuration_Reference.md
… docs/Local_Model_Integration_UI_Guide.md
… package.json
… src/common/services/asyncProcessingManager.js
… src/common/services/asyncWorker.js
… src/common/services/cachePromotionManager.js
… src/common/services/cacheStatisticsMonitor.js
… src/common/services/compatibilityLayer.js
… src/common/services/directMLProvider.js
… src/common/services/documentMetadataManager.js
… src/common/services/dynamicResourceAllocator.js
… src/common/services/errorRecoverySystem.js
… src/common/services/gracefulDegradationManager.js
… src/common/services/hardwareAccelerationManager.js
… src/common/services/hardwareAdaptiveConfig.js
… src/common/services/hardwareDetectionService.js
… src/common/services/hierarchicalEmbeddingCache.js
… src/common/services/intelligentFallbackManager.js
… src/common/services/ioBindingManager.js
… src/common/services/llmwareIntegrationService.js
… src/common/services/namedPipeBridge.js
… src/common/services/onnxRuntimeOptimizer.js
… src/common/services/onnxSessionManager.js
… src/common/services/optimizedDocumentProcessor.js
… src/common/services/pluginAPI.js
… src/common/services/pluginSecurity.js
… src/common/services/resourceSharingManager.js
… src/common/services/sessionPoolManager.js